this looks worse than it is
