+++ package.json
@@ -1,29 +1,28 @@
 {
   "name": "doberview-2",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "node app.js"
   },
   "dependencies": {
-    "axios": "^0.21.4",
+    "axios": "0.29.0",
     "bootstrap": "^5.3.3",
     "bootstrap5-toggle": "^4.3.5",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
-    "express": "~4.16.1",
+    "express": "~4.21.1",
     "highcharts": "^10.3.2",
-    "http-errors": "~1.6.3",
+    "http-errors": "~2.0.0",
     "jquery": "^3.6.0",
     "jquery-datetimepicker": "^2.5.21",
-    "jsoneditor": "^9.9.2",
+    "jsoneditor": "^10.1.1",
     "monk": "^7.3.4",
-    "morgan": "~1.9.1",
-    "passport": "^0.5.2",
+    "passport": "^0.7.0",
     "passport-github2": "^0.1.12",
     "pug": "^3.0.2",
     "tabulator-tables": "^6.2.0",
-    "zeromq": "^5.3.1"
+    "zeromq": "^6.1.2"
   },
   "devDependencies": {
     "express-session": "^1.17.2",
