+++ public/javascripts/hosts_scripts.js
