+++ public/javascripts/logs_scripts.js
