+++ public/javascripts/shifts_scripts.js
