+++ public/stylesheets/default_style.css
@@ -65,8 +65,6 @@ ul.CTAs li a {
   display: block;
 }
 
-
-
 @media (max-width: 992px) {
   .navbar-nav .nav-item {
     padding: 0.5rem 0;
@@ -103,7 +101,6 @@ ul.CTAs li a {
   left: 50%;
   text-align: center;
 }
-
 .nav-item{
   white-space: nowrap;
   padding-right: 20px;
