+++ routes/shifts.js
@@ -7,7 +7,7 @@ const config = require("../config/config");
 function ensureAuthenticated(req, res, next) {
   if (!config.use_authentication) { return next(); }
   if (req.isAuthenticated()) { return next(); }
-  res.redirect('/devices?notify_msg=You must be logged in to do that&notify_status=error');
+  res.redirect('/sensors?notify_msg=You must be logged in to do that&notify_status=error');
 }
 
 router.get('/', ensureAuthenticated, function(req, res) {
