+++ views/hosts.pug
@@ -1,6 +1,6 @@
 extends layout
 block extrahead
-  script(src="javascripts/host_scripts.js" type="text/javascript")
+  script(src="javascripts/hosts_scripts.js" type="text/javascript")
   link(rel="stylesheet" href="stylesheets/hosts.css")
 
 block content
