--- views/index.pug
@@ -1,31 +1,0 @@
-doctype html
-html
-  head
-    title=DOBERview
-    link(rel="apple-touch-icon" sizes="180x180" href="/images/apple-touch-icon.png")
-    link(rel="icon" type="image/png" href="/images/favicon-196x196.png" sizes="196x196")
-    link(rel="icon" type="image/png" href="/images/favicon-96x96.png" sizes="96x96")
-    link(rel="icon" type="image/png" href="/images/favicon-32x32.png" sizes="32x32")
-    link(rel="icon" type="image/png" href="/images/favicon-16x16.png" sizes="16x16")
-    link(rel="icon" type="image/png" href="/images/favicon-128.png" sizes="128x128")
-
-
-    link(rel="stylesheet" href="/modules/bootstrap/dist/css/bootstrap.css")
-
-    script(src="/modules/jquery/dist/jquery.js" type="text/javascript")
-    script(src="/modules/bootstrap/dist/js/bootstrap.js" type="text/javascript")
-    script(src="https://kit.fontawesome.com/fd56c7481d.js" crossorigin="anonymous" type="text/javascript")
-    script(src="javascripts/default_scripts.js" type="text/javascript")
-
-    meta(name="application-name" content="pagetitle")
-
-  body
-      h1.display-1.text-center Welcome to DoberView!
-      h1.text-center Please select an experiment
-      br
-      .d-grid.gap-2.col-6.mx-auto
-        each real_name, display_name in experiments
-          button.btn.btn-large.btn-primary(type='button' onclick=`SetExperiment({'name':'${real_name}'}, {})`) #{display_name}
-
-      script.
-        $(document).ready(function () {});
