+++ views/layout.pug
@@ -31,13 +31,13 @@ html(data-bs-theme="dark")
           img#exp_logo(src="/images/pancake.png" alt="Pancake" width="40" style="margin-top:15px")
         ul.list-unstyled.components
           li#loberview(class="colored")
-            a(href="/systems" hidden=hide_systems)
+            a(href="/overview" hidden=hide_systems)
               i.fas.fa-object-group
               span Overview
-            a(href="/devices")
+            a(href="/sensors")
               i.fa-solid.fa-table-list
               span Sensors
-            a(href='/pipeline')
+            a(href='/pipelines')
               i.fas.fa-code-branch
               span Pipelines
             a(href='/hosts' hidden=hide_hosts)
