+++ views/logs.pug
@@ -6,7 +6,7 @@ block extrahead
   link(rel="stylesheet" href="/stylesheets/logs.css")
   script(type="text/javascript" src="/modules/tabulator-tables/dist/js/tabulator.js")
   script(type="text/javascript" src="/modules/jquery-datetimepicker/build/jquery.datetimepicker.full.min.js")
-  script(type="text/javascript" src="javascripts/logs.js")
+  script(type="text/javascript" src="javascripts/logs_scripts.js")
 
 block content
   div.main-container
