+++ views/overview.pug
@@ -1,7 +1,7 @@
 extends layout
 
 block extrahead
-  script(src="/javascripts/systems.js" type="text/javascript")
+  script(src="/javascripts/overview_scripts.js" type="text/javascript")
   //link(rel='stylesheet' href="/stylesheets/systems.css")
 
 block content
