+++ views/sensors.pug
@@ -1,7 +1,7 @@
 extends layout
 
 block extrahead
-  script(src="javascripts/full_system.js", type="text/javascript")
+  script(src="javascripts/sensors_scripts.js", type="text/javascript")
 
 block content
   div.main-container
