+++ views/shifts.pug
@@ -2,7 +2,7 @@ extends layout
 
 block extrahead
   link(rel="stylesheet" href="/stylesheets/shifts.css")
-  script(src="/javascripts/shifts.js" type="text/javascript")
+  script(src="/javascripts/shifts_scripts.js" type="text/javascript")
 
 
 block content
